requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.59.5
google-generativeai==0.8.3
nltk==3.9.1
python-multipart==0.0.18
jinja2==3.1.5